+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.2] - 2019-10-16
 - allow using variables in template file/dir names
 - provide pluggable Javascript (and other) package manager support
 - simple working syncplugin support
 - improve frontend plugin creation
+- generate frontend plugin names from entry point group by stemming
 
 ## [0.4.0] - 2019-10-12
 - Change API to an easier one, following Marty Alchin's Simple Plugin Framework